moved counters into separated folders, snapshots into separated files
